cleanup; saner deleted logik; other wallpaper set/change support;
this version removes most of the old and unused code.

deleted files will now stay in the database but get an additional delete flag,
this allows deleted files to be remembered.

the tool to generate and set wallpapers is now configurable,
this enables the program to be used on linux